Changes to register and Branch details
Some areas missed (postcode and incorrect email fields - now changed).
Also I have set the Branch of a new company to be that of the company
with a branch name of 'Head office'.  See notes in ICompanyBusiness that
I have put as I'm not sure this is the best way.  It works but maybe it
is not the correct way of getting this detail (i.e. my if..else.....you
will see it.  If this is not the way to do it just let me know the right
way so I know from hereon out.  Thanks